+++ docs/app/[lng]/(main)/docs/page.tsx
@@ -1,5 +1,14 @@
+'use client'
+
+import { useEffect } from 'react'
 import dynamic from 'next/dynamic'
 import { Container } from '@mui/material'
+import Prism from 'prismjs'
+
+// Load the languages to highlight
+import 'prismjs/components/prism-bash'
+import 'prismjs/components/prism-jsx'
+import 'prismjs/components/prism-tsx'
 
 // Types
 import type { PageProps } from '@/types/app/pages'
@@ -16,6 +25,11 @@ const HomePage = (props: PageProps) => {
 
 	// Variables
 	const lng = params?.lng ?? LocaleConfig.default
+
+	// Callbacks
+	useEffect(() => {
+		Prism.highlightAll()
+	}, [])
 
 	return (
 		<Container maxWidth="xl">