+++ docs/types/app/pages/index.d.ts
@@ -1,6 +1,1 @@
-// Types
-import type { ReactNode } from 'react'
-
-export type PageProps = {
-	params?: { lng: string }
-}
+export type PageProps = { params?: { lng: string } }