+++ docs/types/components/atoms/code.d.ts
@@ -1,0 +1,7 @@
+// Types
+import type { ReactNode } from 'react'
+
+export type CodeProps = {
+	className?: string
+	children?: ReactNode
+}