+++ docs/types/components/molecules/list.d.ts
@@ -1,0 +1,10 @@
+export type ListItemProp = {
+	title: string
+	link: string
+	children?: ListItem[]
+}
+
+export type ListProps = {
+	lng: string
+	items: ListItemProp[]
+}