+++ docs/types/components/organisms/drawer.d.ts
@@ -1,0 +1,1 @@
+export type DrawerProps = { lng: string }